Only keep :latest tag for local images
Images are built, tagged with registry and content-addressable tag,
pushed, re-tagged as :latest, and the original tag is removed.

This prevents Cog from using a ton of disk space, but builds are still
fast since the cached layers from the last build are still kept around
thanks to the :latest tag.

Signed-off-by: andreasjansson <andreas@replicate.ai>