Display better error when output type is undefined
If an output type isn't specified, Pydantic expects the model to return
nothing. When the model returns an output and it's compared to `None`,
an error is logged and a 500 error is returned.

This change explicitly raises an error when the output type is undefined
and raises an error message, rather than defaulting to `None`. The error
message provides explanation of how to fix the issue more clearly than
the original behaviour.

Signed-off-by: Dominic Baggott <dominic.baggott@gmail.com>