Send response when queue worker starts predict
Currently the only way to know whether Cog has picked up a request is to
watch for jobs to be claimed from the request queue. Sending an empty
processing message is a much cleaner way to handle that, and will be
picked up by whatever system folks are using to listen to non terminal
responses.

Signed-off-by: Dominic Baggott <dominic.baggott@gmail.com>