Add an explicit handler to die on SIGTERM
If SIGTERM arrives before we have started RedisQueueWorker (where we
explicitly register a handler for that signal) and we are running as PID
1, the kernel will decline to kill the process.

This means that there is a period during startup where a SIGTERM will be
ignored.

To avoid/minimise the possibility of this, this commit sets an explicit
SIGTERM handler to quit the process which is used until overridden by
the more graceful handler inside RedisQueueWorker.

Signed-off-by: Nick Stenning <nick@whiteink.com>