Incorporate READY/BUSY status into model healthcheck
This commit tightens up the contract between director and the model
container's /health-check endpoint.

Specifically, the `status` field of the healthcheck endpoint now reports
one of four health statuses: STARTING, READY, BUSY, SETUP_FAILED.

The expected state transitions between these four states are as follows:

                             ┌────────┐
                             ▼        │
            STARTING ────► READY     BUSY
                │            │        ▲
                │            └────────┘
                ▼
          SETUP_FAILED

Once setup is complete, we expect the model pod to cycle between READY
and BUSY. If any other status is seen, or if BUSY is seen when we expect
the container to be READY, we abort the pod.

Signed-off-by: Nick Stenning <nick@whiteink.com>