Run setup on a thread
Currently, Runner.setup() blocks. When we call it in the FastAPI app
"startup" hook, we block the initialisation of the HTTP server itself.

That in turn means that any attempts to connect to the HTTP interface
while setup is running are refused, so we cannot successfully shut down
a model container while setup is running. This leaves model pods in a
broken state for up to 30m (the K8S termination grace period) until they
are killed.

This changes Runner.setup() to a non-blocking method that returns an
AsyncResult. We can then poll this object when the healthcheck endpoint
is called to find out whether setup is complete.

Note that it is not a requirement to call the healthcheck endpoint.
While setup is running, prediction endpoints will return a 409 Conflict
indicating that the predictor is busy, and will start accepting requests
once setup has completed. (Although note that currently, they will still
accept requests even if setup has failed.)

Signed-off-by: Nick Stenning <nick@whiteink.com>